# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Spark X Test Platform API
  version: 0.1.0
  description: |
    MVP API baseline for the independent automation test platform.
    The first internal release has no user authentication, but is restricted to
    the controlled intranet. Worker transport is handled by the queue and is not
    part of this public/control-plane API.
servers:
  - url: /api/v1
...
paths:
...
  /systems:
    get:
      tags: [Systems]
      operationId: listSystems
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Systems page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemPage"
    post:
      tags: [Systems]
      operationId: createSystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemInput"
      responses:
        "201":
          description: System created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "400":
          $ref: "#/components/responses/BadRequest"
  /systems/{systemId}:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    get:
      tags: [Systems]
      operationId: getSystem
      responses:
        "200":
          description: System details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Systems]
      operationId: updateSystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemPatch"
      responses:
        "200":
          description: System updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /systems/{systemId}/modules:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    get:
      tags: [Systems]
      operationId: listModules
      responses:
        "200":
          description: Modules in display order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      tags: [Systems]
      operationId: createModule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Module created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          $ref: "#/components/responses/BadRequest"
  /systems/{systemId}/environments:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    get:
      tags: [Environments]
      operationId: listEnvironments
      responses:
        "200":
          description: Environments for a system.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Environment"
    post:
      tags: [Environments]
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvironmentInput"
      responses:
        "201":
          description: Environment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "400":
          $ref: "#/components/responses/BadRequest"
  /environments/{environmentId}:
    parameters:
      - $ref: "#/components/parameters/EnvironmentId"
    get:
      tags: [Environments]
      operationId: getEnvironment
      responses:
        "200":
          description: Environment details without secret values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Environments]
      operationId: updateEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvironmentPatch"
      responses:
        "200":
          description: Environment updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /datasets:
    get:
      tags: [Datasets]
      operationId: listDatasets
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: systemId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dataset page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericPage"
    post:
      tags: [Datasets]
      operationId: createDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatasetInput"
      responses:
        "201":
          description: Dataset and initial immutable version created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResource"
...
  /case-templates:
    get:
      tags: [Templates]
      operationId: listCaseTemplates
      parameters:
        - name: systemId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Built-in and adapter-provided case templates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenericResource"
  /shared-steps:
    get:
      tags: [Shared Steps]
      operationId: listSharedSteps
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Shared step page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericPage"
    post:
      tags: [Shared Steps]
      operationId: createSharedStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NamedDefinitionInput"
      responses:
        "201":
          description: Shared step created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResource"
...
  /secrets:
    get:
      tags: [Secrets]
      operationId: listSecretMetadata
      responses:
        "200":
          description: Secret metadata; values are never returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretMetadata"
    post:
      tags: [Secrets]
      operationId: upsertSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretInput"
      responses:
        "200":
          description: Secret stored or rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretMetadata"
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    SystemId:
      name: systemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EnvironmentId:
      name: environmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    SystemInput:
      type: object
      additionalProperties: false
      required: [key, name]
      properties:
        key:
          type: string
          pattern: "^[a-z][a-z0-9-]+$"
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        concurrencyLimit:
          type: integer
          minimum: 1
          maximum: 100
          default: 5
    SystemPatch:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        status:
          enum: [active, archived]
        concurrencyLimit:
          type: integer
          minimum: 1
          maximum: 100
    System:
      allOf:
        - $ref: "#/components/schemas/SystemInput"
        - type: object
          required: [id, status, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            status:
              enum: [active, archived]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    SystemPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/System"
        nextCursor:
          type: [string, "null"]
    ModuleInput:
      type: object
      additionalProperties: false
      required: [key, name]
      properties:
        key:
          type: string
          pattern: "^[a-z][a-z0-9-]+$"
        name:
          type: string
          minLength: 1
          maxLength: 200
        sortOrder:
          type: integer
          minimum: 0
          default: 0
    Module:
      allOf:
        - $ref: "#/components/schemas/ModuleInput"
        - type: object
          required: [id, systemId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            systemId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
...
    EnvironmentPatch:
      type: object
      description: Partial EnvironmentInput. Secret values are referenced separately.
      additionalProperties: false
      properties:
        name:
          type: string
        kind:
          enum: [test, staging, production]
        baseUrl:
          type: string
          format: uri
        actionLevel:
          enum: [read, write, dangerous]
        allowlist:
          type: array
          minItems: 1
          items:
            type: object
        timezone:
          type: string
        concurrencyLimit:
          type: integer
          minimum: 1
          maximum: 100
        adapterKey:
          type: string
        adapterConfig:
          type: object
        status:
          enum: [active, disabled, archived]
    Environment:
      allOf:
        - $ref: "#/components/schemas/EnvironmentInput"
        - type: object
          required: [id, systemId, status, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            systemId:
              type: string
              format: uuid
            status:
              enum: [active, disabled, archived]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    GenericPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GenericResource"
        nextCursor:
          type: [string, "null"]
...